move client errors to parent client package
